Bug fixes for file extensions and 0 bins.

+++ src/itermag/workflow/rules/map_reads_mags.smk
@@ -15,29 +15,33 @@ rule extract_reads:
         
         cat iter_{wildcards.iteration}/binning_filtered/*.fa > iter_{wildcards.iteration}/bin_mapping/all_mag_contigs.fna 2>/dev/null || touch iter_{wildcards.iteration}/bin_mapping/all_mag_contigs.fna
 
-        coverm make -r iter_{wildcards.iteration}/bin_mapping/all_mag_contigs.fna \
-            -1 {config[forward]} \
-            -2 {config[reverse]} \
-            -o iter_{wildcards.iteration}/bin_mapping/coverm
-        
-        mv iter_{wildcards.iteration}/bin_mapping/coverm/*.bam iter_{wildcards.iteration}/bin_mapping/coverm.bam
-        rm -rf iter_{wildcards.iteration}/bin_mapping/coverm
+        if [ ! -s iter_{wildcards.iteration}/bin_mapping/all_mag_contigs.fna ]; then
+            echo "No high-quality MAGs found. Skipping read mapping."
+            touch {output.bam} {output.reads_unmapped_1} {output.reads_unmapped_2}
+        else
+            coverm make -r iter_{wildcards.iteration}/bin_mapping/all_mag_contigs.fna \
+                -1 {config[forward]} \
+                -2 {config[reverse]} \
+                -o iter_{wildcards.iteration}/bin_mapping/coverm
+            
+            mv iter_{wildcards.iteration}/bin_mapping/coverm/*.bam iter_{wildcards.iteration}/bin_mapping/coverm.bam
+            rm -rf iter_{wildcards.iteration}/bin_mapping/coverm
 
-        coverm filter \
-            -b iter_{wildcards.iteration}/bin_mapping/coverm.bam \
-            -o {output.bam} \
-            --inverse \
-            --min-read-percent-identity-pair 95 \
-            --min-read-aligned-percent-pair 75 \
-            --proper-pairs-only
+            coverm filter \
+                -b iter_{wildcards.iteration}/bin_mapping/coverm.bam \
+                -o {output.bam} \
+                --inverse \
+                --min-read-percent-identity-pair 95 \
+                --min-read-aligned-percent-pair 75 \
+                --proper-pairs-only
 
-        samtools fastq \
-            -@ $(({threads} - 1)) \
-            {output.bam} \
-            -1 {output.reads_unmapped_1} \
-            -2 {output.reads_unmapped_2} \
-            -0 /dev/null \
-            -s /dev/null \
-            -n
-
+            samtools fastq \
+                -@ $(({threads} - 1)) \
+                {output.bam} \
+                -1 {output.reads_unmapped_1} \
+                -2 {output.reads_unmapped_2} \
+                -0 /dev/null \
+                -s /dev/null \
+                -n
+        fi
         """
